Add Autopopulating Tags Page
A new page is added which groups all posts by their tags. This links from the
tags listed on the bottom of each card on the front page. Also added links to
the navbar to both the home and tags pages, and reformatted the existing tags
to be consistent ([tag1, tag2, ...]).